After a couple weeks studying dual quaternions, I am choosing not to
support them in this library. The reasons being is that their use is not that common, and un-needed for the games built
thus far with this library, and there is a lot to learn, test, and think about their implementation details. If one
needs support for dual quaternions, it is best to use another library, as they will likely not ever be supported in
gamebox-math.