Use the previous dependency-tracking implementation
along with a flag that specifies if the compilation of a defun/defun-gpu form should trigger
dependency tracking.

2 exposed functions, enable-dependency-tracking and disable-dependency-tracking can be used to
control when dependency tracking should be enabled. This is important because dependency tracking is
an expensive process, and probably should not be done at asd load time. It is preferable to enable
dependency tracking by called (enable-dependency-tracking) sometime after asdf has loaded the
system. Dependency tracking is really only to make live recompilation of shader functions possible,
and therefor only needs to be enabled when a game project is running after compilation.