Disable pseudo-Lisp functions that are defined along with GPU functions.
This was a hack to get editing signature hints and allow jumping to
definition, but it has problems with both of these things it tries to
solve, as well as other issues.

For editor hints, it will always show the signature of the last defined
GPU function of the same name. Shadow allows function overloading, with
different signatures for the same name, so the pseudo-Lisp function is
completely useless and misleading in this case.

For jump to definition, the same problem occurs. It may not jump to the
actual function written, and instead may jump to the last function
defined with the same name. This is very unintuitive and could lead to a
bit of lost development time by a user.

Additionally, pseudo-Lisp functions being defined will raise errors if
they conflict with existing actual Lisp functions with the same name.

Finally, it's my opinion that top-level functions should not be created
in the package namespace behind a user's back. This is just my opinion,
but a strong one, and one of many reasons why I avoid DEFSTRUCT a lot.

For these reasons, pseudo-Lisp functions have been removed.