Add function: rebind-blocks
This should be called from user code in a hook that is called when a GPU
function is modified. This will rebind all blocks relevant to the
programs affected by the compilation of the function. This is needed
because shader blocks are unbound by the driver when a program is compiled.