Always track dependencies at load time.
Before this commit, there was an experiemental feature that disabled
dependency tracking at load time. This was becaus the Varjo compiler is
doing extra work to type check functions at they are defined, and does
so for each type of OpenGL shader stage, so that it can gather
information for error reporting and catch mistakes when they are defined
instead of when they are used. This feature introduced a rather
unintuitive side effect in order to significantly reduce CL load time -
non-stage functions (those while are called from a stage function) and
children thereof, cannot be re-compiled for live coding - no result will
be seen, with no indication as to why. The reason why is simple. A stage
function must be compiled before the rest of the pipeline can track
dependencies. The solution would have been to ensure that you always
redefine the stage function before you modify a helper function, but
this would ensue lots of confusion, both by myself, and any other users
of Shadow. For this reason, dependency tracking is enable, as well as
the expensive type-checking of shader functions at CL load time. Since
it isn't very often that a CL image is restarted, and it is only of
concern with large collections of shader functions, such as those
contained in the Umbra project, this seemed like the correct thing to
do, rather than insert some fine print to be aware of, that even I, the
developer, will forget in some months time. I'd rather this be useful
and intuitive than to jump through hoops to get what you expect to
happen to actually occur.