Remove experimental enabling/disabling of dependency tracking that was never exported to the public API.
This resulted in bad GLSL not detected at compile time, and also the bad
effect of recompilation of functions not possible until a parent
function has been recompiled.

An unavoidable side-effect is since dependency tracking is no longer
disabled during Lisp file loading/compiling, it takes significantly
longer to compile shader functions at compile time now, since varjo has
to run it semantic analyzer for each function for every possible shader
stage that function can be a part of. This is not a big deal, and the
benefits of detecting shader problems early far outweigh the increase in
image startup time.