Fixed matrix constructors. Most (matN matM) constructors got row/column majorness wrong resulting in transposed matrix results.
For example, a copy of a mat4 can be created just like in GLSL by doing:

(m4:mat some-mat4)

That is, constructing a matrix given an existing matrix of the same
type, results in a copy. However, it wasn't writing the components in
the correct order and this did not result in a copy, but a transposed matrix.