Add a scaling factor for Perlin 3D
KdotJPG, the OpenSimplex author commented on the code, saying that the
proper scale for 3D Perlin is 0.9649214285521897.

He also mentions a scaling factor for 4D Perlin, which was tested to be
plausible, but I'd like to run some more tests and/or see if Perlin can
be made to be more consistent, before commiting to that change, since
reaching the maximum range of -/+ ~1.5 seems incredibly rare, and would
mute values by a lot in almost all cases.