Don't make Perlin the default generator for fractals.
This was recommended by KdotJPG, and I absolutely couldn't agree more
with them.

"Perlin being considered to be the "default" or "go-to" coherent noise
choice, is a problematic status quo in the procedural generation
community. It causes people to use it in projects where its visible
squareness problems are an entirely unnecessary compromise."